Move change mote type to GUI
Put the change mote type dialog
in the GUI class, and update
the XML in the same way as the
other simulation reconfiguration.
