+++ java/org/contikios/cooja/mote/BaseContikiMoteType.java
@@ -206,23 +206,16 @@ public abstract class BaseContikiMoteType extends AbstractApplicationMoteType {
             logger.warn("Can't find mote interface class: " + name);
             return false;
           }
-          var hasInterface = false;
-          for (var moteInterface : getAllMoteInterfaceClasses()) {
-            if (moteInterface == clazz) {
-              hasInterface = true;
-              break;
-            }
-          }
-          // Skip incompatible interfaces. Required for changing mote type when reconfiguring a simulation.
-          if (hasInterface) {
-            moteInterfaceClasses.add(clazz);
-          }
+          moteInterfaceClasses.add(clazz);
         }
         case "contikibasedir", "contikicoredir", "projectdir", "compilefile", "process", "sensor", "coreinterface" -> {
           logger.fatal("Old Cooja mote type detected, aborting..");
           return false;
         }
       }
+    }
+    if (moteInterfaceClasses.isEmpty()) { // Old MspMote simulation, or reconfigured simulation.
+      moteInterfaceClasses.addAll(Arrays.asList(getDefaultMoteInterfaceClasses()));
     }
     return true;
   }
